SC-29859 Remove sonar-logo and add sonar-logo-white
