--- src/components/logos/LogoSonar.tsx
@@ -1,120 +1,0 @@
-/*
- * Echoes React
- * Copyright (C) 2023-2025 SonarSource SA
- * mailto:info AT sonarsource DOT com
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- */
-
-import { forwardRef } from 'react';
-import { useIntl } from 'react-intl';
-import { LogoProps, LogoSvgWrapper } from './logoUtils';
-
-import { cssVar } from '~utils/design-tokens';
-
-export const LogoSonar = forwardRef<SVGSVGElement, LogoProps>(
-  ({ hasText = false, size, ...props }, ref) => {
-    const intl = useIntl();
-
-    return hasText ? (
-      <LogoSvgWrapper
-        fill="none"
-        ref={ref}
-        role="img"
-        size={size}
-        viewBox="0 0 152 48"
-        xmlns="http://www.w3.org/2000/svg"
-        {...props}>
-        <title>
-          {intl.formatMessage({
-            id: 'logo.sonar',
-            defaultMessage: 'Sonar',
-            description: 'Alternative text for the Sonar company logo',
-          })}
-        </title>
-        <path
-          d="M40.6566 20.8663C40.6566 16.9701 39.9499 13.2227 38.5458 9.73572C38.2669 9.8566 38.0437 9.96819 37.8949 10.0426C37.8949 10.0426 37.8949 10.0426 37.8856 10.0426C37.244 10.368 36.8907 11.0654 36.9743 11.7721C37.1417 13.2227 37.244 14.9151 37.244 16.4401C37.244 23.5722 34.9379 30.4068 30.7628 35.6977C26.8108 40.7005 21.5106 43.9085 15.7454 44.8291C15.764 44.8291 15.7733 44.8477 15.7919 44.857C19.1673 46.7818 23.2959 46.9027 26.7736 45.1917C34.9193 41.184 40.6566 31.7923 40.6566 20.8663Z"
-          fill={cssVar('logos-colors-sonar-whale')}
-        />
-        <path
-          d="M42.2465 9.01042C41.7072 8.99182 41.2051 9.02901 40.7308 9.1034C42.1814 12.8043 42.9253 16.7562 42.9253 20.8663C42.9253 33.1127 36.1001 43.9922 26.1318 47.9535C26.634 47.9813 27.1454 48 27.6568 48C27.6847 48 27.7033 48 27.7312 48C32.5294 47.9814 37.0857 45.7869 40.0985 42.0581C43.911 37.3343 46.2543 30.974 46.2543 24C46.2543 18.5974 44.8688 13.4273 42.2465 9.01042Z"
-          fill={cssVar('logos-colors-sonar-whale')}
-        />
-        <path
-          d="M51.657 24C51.657 20.7361 51.0061 17.6304 49.8251 14.7942C49.2858 13.4924 48.5233 12.2836 47.519 11.2979C46.9611 10.74 46.3567 10.2751 45.7058 9.94031C45.5384 9.84733 45.3617 9.77294 45.1944 9.69855C47.3703 14.0318 48.514 18.9508 48.514 24C48.514 31.0856 46.2079 37.9202 42.1072 43.1553C47.9003 38.7757 51.6477 31.8202 51.6477 24H51.657Z"
-          fill={cssVar('logos-colors-sonar-whale')}
-        />
-        <path
-          d="M44.8781 7.2809C40.517 2.78962 34.4171 0 27.6569 0C14.3876 0 3.51741 10.9911 3.65689 24.2511C3.69409 27.4963 4.3636 30.5835 5.56313 33.401C5.17259 33.3452 4.77274 33.3266 4.3729 33.3452C3.29425 33.3917 1.77855 33.0105 1.09045 32.8152C0.420938 32.6292 -0.211373 33.1778 0.0675883 33.8195C0.0675883 33.8195 0.0675883 33.8287 0.0675883 33.838C2.44806 39.2964 7.73903 42.8485 12.7789 42.8392C12.7789 42.8392 12.7789 42.8392 12.7882 42.8392C13.123 42.8392 13.4577 42.8206 13.7925 42.7834C25.6019 41.8814 34.975 30.4161 34.975 16.4494C34.975 15.5661 34.9378 14.692 34.8634 13.8179C34.8262 13.3437 34.2125 13.167 33.9335 13.5482C30.6697 17.8721 25.825 25.4692 19.7716 21.5079C17.4469 19.9922 16.2009 16.7191 16.8983 13.3902C18.1629 7.41108 25.1276 3.48702 28.1962 4.71445C28.438 4.80744 29.3028 5.15149 29.591 5.94188C30.0374 7.14142 28.6891 8.13638 28.0009 10.2658C27.257 12.5719 27.6755 15.0081 28.1962 15.1197C28.7727 15.2406 29.7584 13.1019 32.1668 11.1585C33.9149 9.74506 35.2353 9.35451 36.0629 7.85742C36.8626 6.41612 35.4399 4.86323 35.9235 4.72375C36.4163 4.58427 36.8626 6.26734 38.3039 7.10422C39.5128 7.80163 40.5821 7.36459 42.1629 7.2995C42.1629 7.2995 42.8417 7.2716 43.9204 7.42968C43.9204 7.42968 44.2272 7.47617 44.72 7.57846C44.8967 7.61565 45.0083 7.40178 44.8874 7.2809H44.8781ZM26.9874 25.7854C27.1734 25.5808 27.5081 25.1902 28.0753 25.1251C28.1869 25.1158 28.7634 25.0601 29.1447 25.432C29.7119 25.9713 29.5538 27.0128 29.0517 27.6172C28.5868 28.1751 27.6662 28.5378 27.0246 28.1193C26.4853 27.766 26.4574 27.0314 26.4574 26.9663C26.4574 26.3898 26.7921 25.9992 26.9688 25.7854H26.9874Z"
-          fill={cssVar('logos-colors-text')}
-        />
-        <path
-          d="M65.3353 35.2421C64.1544 34.7307 63.2338 33.9961 62.555 33.0476C61.9599 32.2108 61.6251 31.253 61.56 30.165C61.5414 29.9326 61.746 29.7373 61.9785 29.7373H63.5313C64.4519 29.7373 65.3446 30.2022 65.7444 31.0298C65.9304 31.4204 66.1908 31.7644 66.5255 32.0713C67.1857 32.685 68.1156 32.9826 69.3152 32.9826C70.5147 32.9826 71.5189 32.685 72.2163 32.0806C72.9138 31.4855 73.2578 30.7044 73.2578 29.7652C73.2578 29.0306 73.0439 28.4262 72.6162 27.9705C72.1792 27.5056 71.6491 27.1523 71.0075 26.9012C70.3659 26.6501 69.4825 26.3712 68.3388 26.0829C66.9068 25.7017 65.7444 25.3204 64.8518 24.9299C63.9591 24.5393 63.1966 23.9349 62.5643 23.1166C61.932 22.2983 61.6158 21.2011 61.6158 19.8342C61.6158 18.5695 61.932 17.4723 62.5643 16.5238C63.1966 15.5754 64.08 14.8501 65.2144 14.3479C66.3489 13.8458 67.6693 13.5947 69.1571 13.5947C71.2772 13.5947 73.0253 14.1248 74.383 15.1848C75.629 16.1612 76.3636 17.463 76.5868 19.0903C76.6239 19.332 76.4194 19.5552 76.1776 19.5552H73.016C72.8115 19.5552 72.6534 19.4064 72.6162 19.2019C72.486 18.5509 72.1234 17.9837 71.5282 17.5002C70.8308 16.933 69.9196 16.6447 68.7851 16.6447C67.753 16.6447 66.9161 16.9051 66.2652 17.4351C65.6143 17.9651 65.2888 18.7183 65.2888 19.704C65.2888 20.3735 65.4934 20.9314 65.9025 21.3592C66.3117 21.7869 66.8324 22.131 67.4647 22.382C68.097 22.6331 68.9525 22.912 70.0498 23.2003C71.5004 23.6001 72.6813 24 73.6019 24.3998C74.5224 24.7997 75.2942 25.4134 75.9358 26.241C76.5774 27.0686 76.8936 28.1844 76.8936 29.5699C76.8936 30.6858 76.596 31.7365 75.9916 32.7222C75.3872 33.7079 74.5224 34.5075 73.3694 35.1027C72.2257 35.7071 70.868 36.0046 69.3152 36.0046C67.846 36.0046 66.5162 35.7443 65.3446 35.2329L65.3353 35.2421Z"
-          fill={cssVar('logos-colors-text')}
-        />
-        <path
-          d="M83.7561 34.9632C82.4264 34.2193 81.3943 33.1685 80.6318 31.8109C79.8786 30.4533 79.4973 28.8818 79.4973 27.0965C79.4973 25.3111 79.8879 23.7582 80.669 22.3913C81.45 21.0244 82.5101 19.9737 83.8584 19.2391C85.2067 18.5045 86.7131 18.1325 88.3683 18.1325C90.0235 18.1325 91.5299 18.5045 92.8782 19.2391C94.2265 19.9737 95.2865 21.0244 96.0676 22.3913C96.8487 23.7582 97.2393 25.3297 97.2393 27.0965C97.2393 28.8632 96.8394 30.4254 96.0397 31.7923C95.2401 33.1592 94.1521 34.2193 92.7759 34.9632C91.3997 35.7071 89.8747 36.079 88.2195 36.079C86.5643 36.079 85.0951 35.7071 83.7747 34.9632H83.7561ZM90.8139 32.2945C91.6228 31.8574 92.2831 31.1879 92.7852 30.3045C93.2873 29.4212 93.5384 28.3518 93.5384 27.0872C93.5384 25.8226 93.2966 24.7532 92.8131 23.8884C92.3296 23.0143 91.6879 22.3541 90.8882 21.9171C90.0886 21.4801 89.2238 21.2569 88.3032 21.2569C87.3826 21.2569 86.5178 21.4801 85.7368 21.9171C84.9464 22.3541 84.3234 23.0143 83.8584 23.8884C83.3935 24.7625 83.161 25.8318 83.161 27.0872C83.161 28.9562 83.6352 30.4068 84.593 31.4297C85.5508 32.4525 86.7503 32.964 88.2009 32.964C89.1215 32.964 89.9956 32.7408 90.8046 32.3038L90.8139 32.2945Z"
-          fill={cssVar('logos-colors-text')}
-        />
-        <path
-          d="M120.477 22.382C121.202 21.0337 122.188 19.9923 123.443 19.2391C124.698 18.4859 126.075 18.1232 127.59 18.1232C128.957 18.1232 130.147 18.3929 131.17 18.9322C132.193 19.4715 133.002 20.1411 133.616 20.9314V20.8664C133.616 19.5087 134.722 18.4022 136.08 18.4022H136.833C137.056 18.4022 137.242 18.5882 137.242 18.8113V35.3724C137.242 35.5955 137.056 35.7815 136.833 35.7815H134.025C133.802 35.7815 133.616 35.5955 133.616 35.3724V33.1965C133.002 34.0147 132.175 34.7029 131.124 35.2515C130.073 35.8001 128.873 36.0698 127.525 36.0698C126.028 36.0698 124.661 35.6885 123.424 34.9167C122.188 34.1449 121.202 33.0756 120.477 31.6994C119.751 30.3232 119.389 28.761 119.389 27.0128C119.389 25.2646 119.751 23.7211 120.477 22.3727V22.382ZM132.881 23.9628C132.388 23.0794 131.737 22.4099 130.938 21.945C130.138 21.4801 129.273 21.2476 128.353 21.2476C127.432 21.2476 126.567 21.4708 125.768 21.9264C124.968 22.382 124.317 23.0423 123.824 23.9163C123.331 24.7904 123.08 25.8226 123.08 27.0221C123.08 28.2216 123.331 29.2724 123.824 30.1744C124.317 31.0763 124.968 31.7645 125.777 32.2387C126.586 32.7129 127.441 32.9454 128.343 32.9454C129.245 32.9454 130.129 32.7129 130.928 32.248C131.728 31.783 132.37 31.1042 132.872 30.2116C133.365 29.3189 133.616 28.2681 133.616 27.0686C133.616 25.8691 133.365 24.8276 132.872 23.9442L132.881 23.9628Z"
-          fill={cssVar('logos-colors-text')}
-        />
-        <path
-          d="M147.908 18.8671C148.67 18.4301 149.572 18.1883 150.586 18.1325C150.818 18.1232 151.014 18.3185 151.014 18.551V21.852H150.102C148.698 21.852 147.629 22.2147 146.904 22.9214C146.178 23.6374 145.816 24.8741 145.816 26.6409V35.3817C145.816 35.6048 145.63 35.7908 145.407 35.7908H142.626C142.403 35.7908 142.217 35.6048 142.217 35.3817V18.8206C142.217 18.5975 142.403 18.4115 142.626 18.4115H143.351C144.709 18.4115 145.816 19.518 145.816 20.8756V20.9407C146.346 20.0574 147.043 19.3693 147.917 18.8764L147.908 18.8671Z"
-          fill={cssVar('logos-colors-text')}
-        />
-        <path
-          d="M115.279 21.4987C114.665 20.3828 113.819 19.5459 112.74 18.9787C111.662 18.4115 110.434 18.1232 109.067 18.1232C108.946 18.1232 108.835 18.1325 108.714 18.1418C108.705 18.1418 108.686 18.1418 108.677 18.1418C108.621 18.1418 108.565 18.1511 108.509 18.1604C108.37 18.1697 108.23 18.1883 108.1 18.2069C107.998 18.2162 107.896 18.2348 107.793 18.2534C107.626 18.2813 107.449 18.3185 107.291 18.3557C107.263 18.3557 107.235 18.3743 107.208 18.3743C106.771 18.4952 106.371 18.6626 105.999 18.8671C105.125 19.36 104.427 20.0481 103.897 20.9314V20.8664C103.897 19.5087 102.791 18.4022 101.433 18.4022H100.708C100.485 18.4022 100.299 18.5882 100.299 18.8113V35.3724C100.299 35.5955 100.485 35.7815 100.708 35.7815H103.488C103.711 35.7815 103.897 35.5955 103.897 35.3724V26.6316C103.897 26.6316 103.897 26.6037 103.897 26.5851V26.0643C103.897 24.5115 104.288 23.3119 105.078 22.4843C105.869 21.6567 106.938 21.2383 108.286 21.2383C109.635 21.2383 110.695 21.6567 111.476 22.4843C112.257 23.3119 112.647 24.5115 112.647 26.0643V35.3724C112.647 35.5955 112.833 35.7815 113.056 35.7815H115.8C116.023 35.7815 116.209 35.5955 116.209 35.3724V25.525C116.209 23.9442 115.902 22.6052 115.297 21.4894L115.279 21.4987Z"
-          fill={cssVar('logos-colors-text')}
-        />
-      </LogoSvgWrapper>
-    ) : (
-      <LogoSvgWrapper
-        fill="none"
-        ref={ref}
-        role="img"
-        size={size}
-        viewBox="0 0 52 48"
-        xmlns="http://www.w3.org/2000/svg"
-        {...props}>
-        <title>
-          {intl.formatMessage({
-            id: 'logo.sonar',
-            defaultMessage: 'Sonar',
-            description: 'Alternative text for the Sonar company logo',
-          })}
-        </title>
-        <path
-          d="M40.6566 20.8663C40.6566 16.9701 39.9499 13.2227 38.5458 9.73572C38.2669 9.8566 38.0437 9.96819 37.8949 10.0426C37.8949 10.0426 37.8949 10.0426 37.8856 10.0426C37.244 10.368 36.8907 11.0654 36.9743 11.7721C37.1417 13.2227 37.244 14.9151 37.244 16.4401C37.244 23.5722 34.9379 30.4068 30.7628 35.6977C26.8108 40.7005 21.5106 43.9085 15.7454 44.8291C15.764 44.8291 15.7733 44.8477 15.7919 44.857C19.1673 46.7818 23.2959 46.9027 26.7736 45.1917C34.9193 41.184 40.6566 31.7923 40.6566 20.8663Z"
-          fill={cssVar('logos-colors-sonar-whale')}
-        />
-        <path
-          d="M42.2465 9.01042C41.7072 8.99182 41.2051 9.02901 40.7308 9.1034C42.1814 12.8043 42.9253 16.7562 42.9253 20.8663C42.9253 33.1127 36.1001 43.9922 26.1318 47.9535C26.634 47.9813 27.1454 48 27.6568 48C27.6847 48 27.7033 48 27.7312 48C32.5294 47.9814 37.0857 45.7869 40.0985 42.0581C43.911 37.3343 46.2543 30.974 46.2543 24C46.2543 18.5974 44.8688 13.4273 42.2465 9.01042Z"
-          fill={cssVar('logos-colors-sonar-whale')}
-        />
-        <path
-          d="M51.657 24C51.657 20.7361 51.0061 17.6304 49.8251 14.7942C49.2858 13.4924 48.5233 12.2836 47.519 11.2979C46.9611 10.74 46.3567 10.2751 45.7058 9.94031C45.5384 9.84733 45.3617 9.77294 45.1944 9.69855C47.3703 14.0318 48.514 18.9508 48.514 24C48.514 31.0856 46.2079 37.9202 42.1072 43.1553C47.9003 38.7757 51.6477 31.8202 51.6477 24H51.657Z"
-          fill={cssVar('logos-colors-sonar-whale')}
-        />
-        <path
-          d="M44.8781 7.2809C40.517 2.78962 34.4171 0 27.6569 0C14.3876 0 3.51741 10.9911 3.65689 24.2511C3.69409 27.4963 4.3636 30.5835 5.56313 33.401C5.17259 33.3452 4.77274 33.3266 4.3729 33.3452C3.29425 33.3917 1.77855 33.0105 1.09045 32.8152C0.420938 32.6292 -0.211373 33.1778 0.0675883 33.8195C0.0675883 33.8195 0.0675883 33.8287 0.0675883 33.838C2.44806 39.2964 7.73903 42.8485 12.7789 42.8392C12.7789 42.8392 12.7789 42.8392 12.7882 42.8392C13.123 42.8392 13.4577 42.8206 13.7925 42.7834C25.6019 41.8814 34.975 30.4161 34.975 16.4494C34.975 15.5661 34.9378 14.692 34.8634 13.8179C34.8262 13.3437 34.2125 13.167 33.9335 13.5482C30.6697 17.8721 25.825 25.4692 19.7716 21.5079C17.4469 19.9922 16.2009 16.7191 16.8983 13.3902C18.1629 7.41108 25.1276 3.48702 28.1962 4.71445C28.438 4.80744 29.3028 5.15149 29.591 5.94188C30.0374 7.14142 28.6891 8.13638 28.0009 10.2658C27.257 12.5719 27.6755 15.0081 28.1962 15.1197C28.7727 15.2406 29.7584 13.1019 32.1668 11.1585C33.9149 9.74506 35.2353 9.35451 36.0629 7.85742C36.8626 6.41612 35.4399 4.86323 35.9235 4.72375C36.4163 4.58427 36.8626 6.26734 38.3039 7.10422C39.5128 7.80163 40.5821 7.36459 42.1629 7.2995C42.1629 7.2995 42.8417 7.2716 43.9204 7.42968C43.9204 7.42968 44.2272 7.47617 44.72 7.57846C44.8967 7.61565 45.0083 7.40178 44.8874 7.2809H44.8781ZM26.9874 25.7854C27.1734 25.5808 27.5081 25.1902 28.0753 25.1251C28.1869 25.1158 28.7634 25.0601 29.1447 25.432C29.7119 25.9713 29.5538 27.0128 29.0517 27.6172C28.5868 28.1751 27.6662 28.5378 27.0246 28.1193C26.4853 27.766 26.4574 27.0314 26.4574 26.9663C26.4574 26.3898 26.7921 25.9992 26.9688 25.7854H26.9874Z"
-          fill={cssVar('logos-colors-text')}
-        />
-      </LogoSvgWrapper>
-    );
-  },
-);
-LogoSonar.displayName = 'LogoSonar';
