+++ src/components/logos/LogoSonarQubeCloudWhite.tsx
@@ -1,0 +1,87 @@
+/*
+ * Echoes React
+ * Copyright (C) 2023-2025 SonarSource SA
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
+import { forwardRef } from 'react';
+import { useIntl } from 'react-intl';
+import { LogoProps, LogoSvgWrapper } from './logoUtils';
+
+import { cssVar } from '~utils/design-tokens';
+
+export const LogoSonarQubeCloudWhite = forwardRef<SVGSVGElement, LogoProps>(
+  ({ size, ...props }, ref) => {
+    const intl = useIntl();
+
+    return (
+      <LogoSvgWrapper
+        fill="none"
+        ref={ref}
+        role="img"
+        size={size}
+        viewBox="0 0 112 36"
+        xmlns="http://www.w3.org/2000/svg"
+        {...props}>
+        <title>
+          {intl.formatMessage({
+            id: 'logo.sonarqubecloudwhite',
+            defaultMessage: 'Sonar',
+            description: 'Alternative text for the Sonar company logo',
+          })}
+        </title>
+        <path
+          d="M30.153 15.6497C30.153 12.7276 29.6289 9.91702 28.5876 7.30176C28.3807 7.39242 28.2151 7.47611 28.1048 7.5319C28.1048 7.5319 28.1048 7.5319 28.0979 7.5319C27.6221 7.77599 27.36 8.29905 27.4221 8.82907C27.5462 9.91702 27.6221 11.1863 27.6221 12.33C27.6221 17.6791 25.9117 22.8051 22.8152 26.7733C19.8843 30.5253 15.9533 32.9314 11.6775 33.6218C11.6913 33.6218 11.6982 33.6357 11.712 33.6427C14.2154 35.0863 17.2774 35.177 19.8567 33.8938C25.8979 30.888 30.153 23.8442 30.153 15.6497Z"
+          fill={cssVar('color-support-white')}
+        />
+        <path
+          d="M31.3324 6.75783C30.9324 6.74388 30.56 6.77177 30.2083 6.82757C31.2841 9.60323 31.8358 12.5672 31.8358 15.6497C31.8358 24.8345 26.7738 32.9942 19.3809 35.9651C19.7533 35.986 20.1326 36 20.5119 36C20.5326 36 20.5464 36 20.567 36C24.1256 35.986 27.5049 34.3402 29.7393 31.5436C32.5669 28.0008 34.3048 23.2305 34.3048 18C34.3048 13.9481 33.2772 10.0705 31.3324 6.75783Z"
+          fill={cssVar('color-support-white')}
+        />
+        <path
+          d="M38.3116 18C38.3116 15.5521 37.8289 13.2228 36.953 11.0957C36.5531 10.1193 35.9875 9.21271 35.2427 8.47346C34.8289 8.05502 34.3807 7.70632 33.8979 7.45525C33.7738 7.38551 33.6428 7.32972 33.5186 7.27393C35.1324 10.5238 35.9806 14.2131 35.9806 18C35.9806 23.3142 34.2703 28.4401 31.229 32.3665C35.5255 29.0818 38.3048 23.8652 38.3048 18H38.3116Z"
+          fill={cssVar('color-support-white')}
+        />
+        <path
+          d="M33.2841 5.46067C30.0496 2.09221 25.5256 0 20.5119 0C10.6706 0 2.60871 8.24332 2.71215 18.1883C2.73974 20.6222 3.23628 22.9376 4.12592 25.0508C3.83627 25.0089 3.53973 24.995 3.24318 25.0089C2.44319 25.0438 1.31907 24.7578 0.808734 24.6114C0.312191 24.4719 -0.156766 24.8834 0.0501272 25.3646C0.0501272 25.3646 0.0501272 25.3716 0.0501272 25.3785C1.81562 29.4723 5.73969 32.1364 9.47756 32.1294C9.47756 32.1294 9.47756 32.1294 9.48445 32.1294C9.73273 32.1294 9.981 32.1155 10.2293 32.0876C18.9877 31.4111 25.9394 22.8121 25.9394 12.3371C25.9394 11.6745 25.9118 11.019 25.8566 10.3634C25.829 10.0077 25.3738 9.87524 25.167 10.1612C22.7463 13.4041 19.1533 19.1019 14.6637 16.131C12.9396 14.9942 12.0154 12.5393 12.5327 10.0426C13.4706 5.55831 18.636 2.61527 20.9119 3.53584C21.0912 3.60558 21.7325 3.86362 21.9463 4.45641C22.2773 5.35606 21.2774 6.10228 20.767 7.69934C20.2153 9.4289 20.5257 11.2561 20.9119 11.3398C21.3394 11.4305 22.0705 9.82642 23.8566 8.36885C25.1532 7.30879 26.1325 7.01589 26.7462 5.89306C27.3393 4.81209 26.2842 3.64742 26.6428 3.54281C27.0083 3.4382 27.3393 4.7005 28.4083 5.32817C29.3048 5.85122 30.0979 5.52344 31.2703 5.47462C31.2703 5.47462 31.7737 5.4537 32.5737 5.57226C32.5737 5.57226 32.8013 5.60713 33.1668 5.68384C33.2979 5.71174 33.3806 5.55134 33.291 5.46067H33.2841ZM20.0153 19.339C20.1532 19.1856 20.4015 18.8927 20.8222 18.8439C20.905 18.8369 21.3325 18.795 21.6153 19.074C22.036 19.4785 21.9187 20.2596 21.5463 20.7129C21.2015 21.1313 20.5188 21.4033 20.0429 21.0895C19.6429 20.8245 19.6222 20.2735 19.6222 20.2247C19.6222 19.7923 19.8705 19.4994 20.0015 19.339H20.0153Z"
+          fill={cssVar('color-support-white')}
+        />
+        <path
+          d="M48.4562 26.4316C47.5804 26.048 46.8976 25.4971 46.3942 24.7857C45.9528 24.1581 45.7045 23.4397 45.6563 22.6238C45.6425 22.4494 45.7942 22.303 45.9666 22.303H47.1183C47.8011 22.303 48.4631 22.6517 48.7597 23.2724C48.8976 23.5653 49.0907 23.8233 49.339 24.0535C49.8286 24.5138 50.5183 24.7369 51.4079 24.7369C52.2975 24.7369 53.0424 24.5137 53.5596 24.0604C54.0768 23.6141 54.332 23.0283 54.332 22.3239C54.332 21.773 54.1734 21.3196 53.8561 20.9779C53.532 20.6292 53.1389 20.3642 52.6631 20.1759C52.1872 19.9876 51.532 19.7784 50.6838 19.5622C49.6217 19.2762 48.7597 18.9903 48.0976 18.6974C47.4355 18.4045 46.87 17.9512 46.4011 17.3375C45.9321 16.7237 45.6977 15.9008 45.6977 14.8756C45.6977 13.9272 45.9321 13.1042 46.4011 12.3929C46.87 11.6815 47.5252 11.1375 48.3666 10.7609C49.2079 10.3843 50.1872 10.196 51.2907 10.196C52.863 10.196 54.1596 10.5936 55.1665 11.3886C56.0906 12.1209 56.6354 13.0972 56.8009 14.3177C56.8285 14.499 56.6768 14.6664 56.4975 14.6664H54.1527C54.001 14.6664 53.8837 14.5548 53.8561 14.4014C53.7596 13.9132 53.4906 13.4878 53.0493 13.1251C52.532 12.6997 51.8562 12.4835 51.0148 12.4835C50.2493 12.4835 49.6286 12.6788 49.1459 13.0763C48.6631 13.4738 48.4217 14.0387 48.4217 14.778C48.4217 15.2801 48.5735 15.6986 48.8769 16.0194C49.1804 16.3402 49.5666 16.5982 50.0355 16.7865C50.5045 16.9748 51.1389 17.184 51.9527 17.4002C53.0286 17.7001 53.9044 18 54.5872 18.2999C55.2699 18.5998 55.8423 19.0601 56.3182 19.6807C56.794 20.3014 57.0285 21.1383 57.0285 22.1774C57.0285 23.0143 56.8078 23.8024 56.3595 24.5416C55.9113 25.2809 55.2699 25.8807 54.4147 26.327C53.5665 26.7803 52.5596 27.0035 51.4079 27.0035C50.3183 27.0035 49.3321 26.8082 48.4631 26.4246L48.4562 26.4316Z"
+          fill={cssVar('color-support-white')}
+        />
+        <path
+          d="M62.1182 26.2224C61.132 25.6645 60.3665 24.8764 59.801 23.8582C59.2423 22.84 58.9596 21.6614 58.9596 20.3223C58.9596 18.9833 59.2492 17.8187 59.8285 16.7935C60.4078 15.7683 61.194 14.9802 62.194 14.4293C63.194 13.8783 64.3112 13.5994 65.5388 13.5994C66.7664 13.5994 67.8836 13.8783 68.8836 14.4293C69.8836 14.9802 70.6698 15.7683 71.2491 16.7935C71.8284 17.8187 72.118 18.9973 72.118 20.3223C72.118 21.6474 71.8215 22.819 71.2284 23.8442C70.6353 24.8694 69.8284 25.6645 68.8077 26.2224C67.787 26.7803 66.656 27.0593 65.4285 27.0593C64.2009 27.0593 63.1112 26.7803 62.132 26.2224H62.1182ZM67.3526 24.2208C67.9525 23.8931 68.4422 23.3909 68.8146 22.7284C69.187 22.0659 69.3732 21.2638 69.3732 20.3154C69.3732 19.3669 69.1939 18.5649 68.8353 17.9163C68.4767 17.2607 68.0008 16.7656 67.4077 16.4378C66.8146 16.11 66.1733 15.9426 65.4905 15.9426C64.8078 15.9426 64.1664 16.11 63.5871 16.4378C63.0009 16.7656 62.5388 17.2607 62.194 17.9163C61.8492 18.5719 61.6768 19.3739 61.6768 20.3154C61.6768 21.7172 62.0285 22.8051 62.7388 23.5722C63.4492 24.3394 64.3388 24.723 65.4147 24.723C66.0974 24.723 66.7457 24.5556 67.3457 24.2278L67.3526 24.2208Z"
+          fill={cssVar('color-support-white')}
+        />
+        <path
+          d="M89.3522 16.7865C89.8901 15.7753 90.6211 14.9942 91.5521 14.4293C92.4831 13.8644 93.5038 13.5924 94.6279 13.5924C95.6417 13.5924 96.5245 13.7947 97.2831 14.1991C98.0417 14.6036 98.6417 15.1058 99.0968 15.6986V15.6497C99.0968 14.6315 99.9175 13.8016 100.924 13.8016H101.483C101.649 13.8016 101.786 13.9411 101.786 14.1085V26.5293C101.786 26.6966 101.649 26.8361 101.483 26.8361H99.4003C99.2348 26.8361 99.0968 26.6966 99.0968 26.5293V24.8973C98.6417 25.511 98.0279 26.0271 97.2486 26.4386C96.4693 26.8501 95.5797 27.0523 94.5797 27.0523C93.4693 27.0523 92.4556 26.7664 91.5383 26.1875C90.6211 25.6087 89.8901 24.8067 89.3522 23.7745C88.8142 22.7423 88.5453 21.5707 88.5453 20.2596C88.5453 18.9485 88.8142 17.7908 89.3522 16.7795V16.7865ZM98.552 17.9721C98.1865 17.3096 97.7038 16.8074 97.1107 16.4587C96.5176 16.11 95.8762 15.9357 95.1935 15.9357C94.5107 15.9357 93.8693 16.1031 93.2762 16.4448C92.6831 16.7865 92.2004 17.2817 91.8349 17.9372C91.4694 18.5928 91.2832 19.3669 91.2832 20.2666C91.2832 21.1662 91.4694 21.9543 91.8349 22.6308C92.2004 23.3072 92.6831 23.8233 93.2831 24.179C93.8831 24.5347 94.5176 24.709 95.1866 24.709C95.8555 24.709 96.5107 24.5347 97.1038 24.186C97.6969 23.8373 98.1727 23.3282 98.5451 22.6587C98.9106 21.9892 99.0968 21.2011 99.0968 20.3014C99.0968 19.4018 98.9106 18.6207 98.5451 17.9582L98.552 17.9721Z"
+          fill={cssVar('color-support-white')}
+        />
+        <path
+          d="M109.697 14.1503C110.262 13.8225 110.931 13.6412 111.683 13.5994C111.855 13.5924 112 13.7388 112 13.9132V16.389H111.324C110.283 16.389 109.49 16.661 108.952 17.191C108.414 17.728 108.145 18.6555 108.145 19.9806V26.5362C108.145 26.7036 108.007 26.8431 107.841 26.8431H105.779C105.614 26.8431 105.476 26.7036 105.476 26.5362V14.1154C105.476 13.9481 105.614 13.8086 105.779 13.8086H106.317C107.324 13.8086 108.145 14.6385 108.145 15.6567V15.7055C108.538 15.043 109.055 14.5269 109.703 14.1573L109.697 14.1503Z"
+          fill={cssVar('color-support-white')}
+        />
+        <path
+          d="M85.497 16.124C85.0418 15.2871 84.4143 14.6594 83.6143 14.234C82.8143 13.8086 81.904 13.5924 80.8902 13.5924C80.8005 13.5924 80.7178 13.5994 80.6281 13.6064C80.6212 13.6064 80.6074 13.6064 80.6005 13.6064C80.5592 13.6064 80.5178 13.6133 80.4764 13.6203C80.373 13.6273 80.2695 13.6412 80.173 13.6552C80.0971 13.6621 80.0212 13.6761 79.9454 13.69C79.8212 13.711 79.6902 13.7389 79.573 13.7668C79.5523 13.7668 79.5316 13.7807 79.5109 13.7807C79.1868 13.8714 78.8902 13.9969 78.6144 14.1503C77.9661 14.52 77.4489 15.036 77.0558 15.6986V15.6497C77.0558 14.6315 76.2351 13.8016 75.2282 13.8016H74.6903C74.5248 13.8016 74.3868 13.9411 74.3868 14.1085V26.5293C74.3868 26.6966 74.5248 26.8361 74.6903 26.8361H76.7523C76.9178 26.8361 77.0558 26.6966 77.0558 26.5293V19.9737C77.0558 19.9737 77.0558 19.9527 77.0558 19.9388V19.5482C77.0558 18.3836 77.3454 17.4839 77.9316 16.8632C78.5178 16.2425 79.3109 15.9287 80.3109 15.9287C81.3109 15.9287 82.0971 16.2425 82.6764 16.8632C83.2557 17.4839 83.5453 18.3836 83.5453 19.5482V26.5293C83.5453 26.6966 83.6832 26.8361 83.8488 26.8361H85.8832C86.0487 26.8361 86.1866 26.6966 86.1866 26.5293V19.1437C86.1866 17.9582 85.9591 16.9539 85.5108 16.117L85.497 16.124Z"
+          fill={cssVar('color-support-white')}
+        />
+      </LogoSvgWrapper>
+    );
+  },
+);
+LogoSonarQubeCloudWhite.displayName = 'LogoSonarQubeCloudWhite';
